+++ mRemoteV1/Config/Connections/Multiuser/ConnectionsUpdateAvailableEventArgs.cs
@@ -1,8 +1,7 @@
 ﻿using System;
 using mRemoteNG.Config.DatabaseConnectors;
 
-
-namespace mRemoteNG.Config.Connections
+namespace mRemoteNG.Config.Connections.Multiuser
 {
     public delegate void ConnectionsUpdateAvailableEventHandler(object sender, ConnectionsUpdateAvailableEventArgs args);
