+++ mRemoteV1/Config/Connections/Multiuser/IConnectionsUpdateChecker.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 
-
-namespace mRemoteNG.Config.Connections
+namespace mRemoteNG.Config.Connections.Multiuser
 {
     public interface IConnectionsUpdateChecker : IDisposable
     {
