+++ mRemoteV1/Config/Connections/Multiuser/SqlConnectionsUpdateChecker.cs
@@ -4,6 +4,7 @@ using System;
 using System.Data;
 using System.Data.SqlClient;
 using System.Threading;
+using mRemoteNG.Config.Connections.Multiuser;
 using mRemoteNG.Config.DatabaseConnectors;
 
 namespace mRemoteNG.Config.Connections
